refactor: move lower-level AWS function to another module
